+++ brewpi-script/scriptlibs/brewpiVersion.py
@@ -29,25 +29,26 @@ def getVersionFromSerial(ser):
     oldTimeOut = ser.timeout
     startTime = time.time()
     if not ser.isOpen():
-        print("Cannot get version from serial port that is not open.")
+        print("\nCannot get version from serial port that is not open.")
 
-    ser.timeout = 1
-    # TODO - Rewrite once Python3 is a given
     try:
-        ser.write('n')  # request version info
-    except TypeError:
-        ser.write(b'n')
+        ser.timeout = 1
+    except SerialException:
+        print("Could not configure port")
+        exit(1)
+
+    ser.write(b'n')
 
     while retries < 10:
         retry = True
-        while 1: # read all lines from serial
+        while 1: # Read all lines from serial
             loopTime = time.time()
             line = None
             try:
                 line = ser.readline()
                 if hasattr(line, 'decode'):
                     line = line.decode(encoding="cp437")
-            except (SerialException) as e:
+            except SerialException as e:
                 pass
             if line:
                 line = asciiToUnicode(line)
@@ -58,10 +59,10 @@ def getVersionFromSerial(ser):
                         retry = False
                         break
             if time.time() - loopTime >= ser.timeout:
-                # have read entire buffer, now just reading data as it comes in. Break to prevent an endless loop.
+                # Have read entire buffer, now just reading data as it comes in. Break to prevent an endless loop
                 break
             if time.time() - startTime >= 10:
-                # try max 10 seconds
+                # Try max 10 seconds
                 retry = False
                 break
 
@@ -75,12 +76,12 @@ def getVersionFromSerial(ser):
             retries += 1
         else:
             break
-    ser.timeout = oldTimeOut # restore previous serial timeout value
+    ser.timeout = oldTimeOut # Restore previous serial timeout value
     return version
 
-
 class AvrInfo:
-    """ Parses and stores the version and other compile-time details reported by the controller """
+    """ Parses and stores the version and other compile-time details reported by the controller. """
+
     version = "v"
     build = "n"
     simulator = "y"
@@ -89,40 +90,41 @@ class AvrInfo:
     log = "l"
     commit = "c"
 
+    shield_diy = "DIY"
     shield_revA = "revA"
     shield_revC = "revC"
     spark_shield_v1 = "V1"
     spark_shield_v2 = "V2"
+    shield_i2c = "I2C"
 
-    shields = {1: shield_revA, 2: shield_revC, 3: spark_shield_v1, 4: spark_shield_v2}
+    shields = {0: shield_diy, 1: shield_revA, 2: shield_revC, 3: spark_shield_v1, 4: spark_shield_v2, 5: shield_i2c}
 
     board_leonardo = "leonardo"
     board_standard = "uno"
     board_mega = "mega"
     board_spark_core = "core"
     board_photon = "photon"
-    board_esp = "esp8266"
+    board_esp8266 = "esp8266"
 
-    boards = {'l': board_leonardo, 's': board_standard, 'm': board_mega, 'x': board_spark_core, 'y': board_photon,
-              'e': board_esp}
+    boards = {'l': board_leonardo, 's': board_standard, 'm': board_mega, 'x': board_spark_core, 'y': board_photon, 'e': board_esp8266}
 
     family_arduino = "Arduino"
     family_spark = "Particle"
-    family_esp = "ESP"
+    family_esp8266 = "ESP8266"
 
     families = { board_leonardo: family_arduino,
                 board_standard: family_arduino,
                 board_mega: family_arduino,
                 board_spark_core: family_spark,
                 board_photon: family_spark,
-                board_esp: family_esp}
+                board_esp8266: family_esp8266}
 
     board_names = { board_leonardo: "Leonardo",
                 board_standard: "Uno",
                 board_mega: "Mega",
                 board_spark_core: "Core",
                 board_photon: "Photon",
-                board_esp: "8266"}
+                board_esp8266: "ESP8266"}
 
     def __init__(self, s=None):
         self.version = LooseVersion("0.0.0")
@@ -148,13 +150,13 @@ class AvrInfo:
         j = None
         try:
             j = json.loads(s)
-        except (json.JSONDecodeError) as e:
+        except json.decoder.JSONDecodeError as e:
             print("JSON decode error: %s" % str(e), file=sys.stderr)
             print("Could not parse version number: " + s, file=sys.stderr)
-        except (UnicodeDecodeError) as e:
+        except UnicodeDecodeError as e:
             print("Unicode decode error: %s" % str(e), file=sys.stderr)
             print("Could not parse version number: " + s, file=sys.stderr)
-        except (TypeError) as e:
+        except TypeError as e:
             print("TypeError: %s" % str(e), file=sys.stderr)
             print("Could not parse version number: " + s, file=sys.stderr)
 
@@ -208,7 +210,7 @@ class AvrInfo:
         if self.shield:
             string += " with a " + str(self.shield) + " shield"
         if(self.simulator):
-           string += ", running as simulator"
+            string += ", running as simulator."
         return string
 
     def isNewer(self, versionString):
@@ -234,4 +236,3 @@ class AvrInfo:
 
     def articleFullName(self):
         return self.article(self.family) + " " + self.fullName()
-