+++ gravity/api/sensors.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from django.shortcuts import render_to_response, redirect
+from django.shortcuts import redirect
 from django.http import JsonResponse
 from django.urls import reverse
 from django.core.exceptions import ObjectDoesNotExist