+++ gravity/models.py
@@ -114,7 +114,7 @@ class GravitySensor(models.Model):
                               help_text='Status of the gravity sensor (used by scripts that interact with it)')
 
     # The beer that is currently active & being logged
-    active_log = models.ForeignKey('GravityLog', null=True, blank=True, default=None,
+    active_log = models.ForeignKey('GravityLog', null=True, blank=True, default=None, on_delete=models.SET_NULL,
                                    help_text='The currently active log of readings')
 
     # The assigned/linked BrewPi device (if applicable)
@@ -508,7 +508,7 @@ class GravityLogPoint(models.Model):
 class TiltTempCalibrationPoint(models.Model):
     TEMP_FORMAT_CHOICES = (('F', 'Fahrenheit'), ('C', 'Celsius'))
 
-    sensor = models.ForeignKey('TiltConfiguration')
+    sensor = models.ForeignKey('TiltConfiguration', on_delete=models.CASCADE)
     orig_value = models.DecimalField(max_digits=8, decimal_places=4, verbose_name="Original (Sensor) Temp Value",
                                      help_text="Original (Sensor) Temp Value")
     actual_value = models.DecimalField(max_digits=8, decimal_places=4, verbose_name="Actual (Measured) Temp Value",
@@ -559,7 +559,7 @@ class TiltTempCalibrationPoint(models.Model):
 
 
 class TiltGravityCalibrationPoint(models.Model):
-    sensor = models.ForeignKey('TiltConfiguration')
+    sensor = models.ForeignKey('TiltConfiguration', on_delete=models.CASCADE)
     actual_gravity = models.DecimalField(max_digits=5, decimal_places=3, verbose_name="Actual (Correct) Gravity value")
     tilt_measured_gravity = models.DecimalField(max_digits=5, decimal_places=3,
                                                 verbose_name="Tilt Measured Gravity Value")
@@ -797,7 +797,7 @@ class TiltBridge(models.Model):
 
 ### iSpindel specific models
 class IspindelGravityCalibrationPoint(models.Model):
-    sensor = models.ForeignKey('IspindelConfiguration')
+    sensor = models.ForeignKey('IspindelConfiguration', on_delete=models.CASCADE)
     angle = models.DecimalField(max_digits=10, decimal_places=7, verbose_name="Angle (Measured by Device)")
     gravity = models.DecimalField(max_digits=8, decimal_places=4, verbose_name="Gravity Value (Measured Manually)")
     created = models.DateTimeField(default=timezone.now)  # So we can track when the calibration was current as of