+++ gravity/tilt/tilt_monitor_aio.py
@@ -1,17 +1,42 @@
 #!/usr/bin/python
 
-# Let's get sentry support going
-from raven import Client
-client = Client('http://3a1cc1f229ae4b0f88a4c6f7b5d8f394:c10eae5fd67a43a58957887a6b2484b1@sentry.optictheory.com:9000/2')
-
-
+# TODO - Figure out the best way to convert this to use sync_to_async calls
 import os, sys
+os.environ["DJANGO_ALLOW_ASYNC_UNSAFE"] = "true"
+
+# Let's get sentry support going
+from sentry_sdk import init, capture_exception
+init('http://3a1cc1f229ae4b0f88a4c6f7b5d8f394:c10eae5fd67a43a58957887a6b2484b1@sentry.optictheory.com:9000/2')
+
 import time, datetime, getopt, pid
 from typing import List, Dict
 import asyncio
-# import argparse, re
-import aioblescan as aiobs
+
+# Initialize logging
 import logging
+LOG = logging.getLogger("tilt")
+LOG.setLevel(logging.INFO)
+
+# We're having environment issues - Check the environment before continuing
+try:
+    import aioblescan as aiobs
+except:
+    LOG.error("Aioblescan not installed - unable to run")
+    exit(1)
+
+try:
+    import pkg_resources
+    from packaging import version
+except:
+    LOG.error("Packaging not installed - unable to run")
+    exit(1)
+
+for package in pkg_resources.working_set:
+    if package.project_name == 'aioblescan':
+        # This is ridiculous but package.parsed_version doesn't return the right type of Version.
+        if version.parse(package.parsed_version.public) < version.parse("0.2.6"):
+            LOG.error("Incorrect aioblescan version installed - unable to run")
+            exit(1)
 
 # done before importing django app as it does setup
 import tilt_monitor_utils
@@ -29,8 +54,7 @@ tilt_monitor_utils.process_monitor_options()
 verbose = tilt_monitor_utils.verbose
 mydev = tilt_monitor_utils.bluetooth_device
 
-LOG = logging.getLogger("tilt")
-LOG.setLevel(logging.INFO)
+
 
 #### The main loop
 
@@ -53,19 +77,18 @@ def processBLEBeacon(data):
     # To make things easier, let's convert the byte string to a hex string first
     if ev.raw_data is None:
         if verbose:
-            print("Event has no raw_data\r\n")
             LOG.error("Event has no raw data")
         return False
 
     raw_data_hex = ev.raw_data.hex()
 
     if len(raw_data_hex) < 80:  # Very quick filter to determine if this is a valid Tilt device
-        if verbose:
-            print("Small raw_data_hex: {}\r\n".format(raw_data_hex))
+        # if verbose:
+        #     LOG.info("Small raw_data_hex: {}".format(raw_data_hex))
         return False
     if "1370f02d74de" not in raw_data_hex:  # Another very quick filter (honestly, might not be faster than just looking at uuid below)
-        if verbose:
-            print("Missing key in raw_data_hex: {}\r\n".format(raw_data_hex))
+        # if verbose:
+        #     LOG.info("Missing key in raw_data_hex: {}".format(raw_data_hex))
         return False
 
     # For testing/viewing raw announcements, uncomment the following
@@ -73,29 +96,38 @@ def processBLEBeacon(data):
     # ev.show(0)
 
     try:
+        mac_addr = ev.retrieve("peer")[0].val
+    except:
+        pass
+
+    try:
         # Let's use some of the functions of aioblesscan to tease out the mfg_specific_data payload
 
-        data = ev.retrieve("Manufacturer Specific Data")
-        payload = data[0].payload
+        manufacturer_data = ev.retrieve("Manufacturer Specific Data")
+        payload = manufacturer_data[0].payload
         payload = payload[1].val.hex()
 
-        # ...and then dissect said payload into a UUID, temp, gravity, and rssi (which isn't actually rssi)
+        # ...and then dissect said payload into a UUID, temp, and gravity
         uuid = payload[4:36]
         temp = int.from_bytes(bytes.fromhex(payload[36:40]), byteorder='big')
         gravity = int.from_bytes(bytes.fromhex(payload[40:44]), byteorder='big')
-        # tx_pwr = int.from_bytes(bytes.fromhex(payload[48:49]), byteorder='big')
-        # rssi = int.from_bytes(bytes.fromhex(payload[49:50]), byteorder='big')
-        rssi = 0  # TODO - Fix this
-    except  Exception as e:
+        # tx_pwr isn't actually tx_pwr on the latest Tilts - I need to figure out what it actually is
+        tx_pwr = int.from_bytes(bytes.fromhex(payload[44:46]), byteorder='big', signed=False)
+        rssi = ev.retrieve("rssi")[-1].val
+
+    except Exception as e:
         LOG.error(e)
-        client.captureException()
+        capture_exception(e)
         return
 
     if verbose:
-        print("Tilt Payload (hex): {}\r\n".format(payload))
+        LOG.info("Tilt Payload (hex): {}".format(raw_data_hex))
 
     color = TiltHydrometer.color_lookup(uuid)  # Map the uuid back to our TiltHydrometer object
     tilts[color].process_decoded_values(gravity, temp, rssi)  # Process the data sent from the Tilt
+
+    #print("Color {} - MAC {}".format(color, mac_addr))
+    #print("Raw Data: `{}`".format(raw_data_hex))
 
     # The Fermentrack specific stuff:
     reload = False
@@ -106,16 +138,24 @@ def processBLEBeacon(data):
 
     for this_tilt in tilts:
         if tilts[this_tilt].should_save():
+            if verbose:
+                LOG.info("Saving {} to Fermentrack".format(this_tilt))
             tilts[this_tilt].save_value_to_fermentrack(verbose=verbose)
 
         if reload:  # Users editing/changing objects in Fermentrack doesn't signal this process so reload on a timer
+            if verbose:
+                LOG.info("Loading {} from Fermentrack".format(this_tilt))
             tilts[this_tilt].load_obj_from_fermentrack()
 
 
 event_loop = asyncio.get_event_loop()
 
 # First create and configure a raw socket
-mysocket = aiobs.create_bt_socket(mydev)
+try:
+    mysocket = aiobs.create_bt_socket(mydev)
+except OSError as e:
+    LOG.error("Unable to create socket - {}".format(e))
+    exit(1)
 
 # create a connection with the raw socket (Uses _create_connection_transport instead of create_connection as this now
 # requires a STREAM socket) - previously was fac=event_loop.create_connection(aiobs.BLEScanRequester,sock=mysocket)
@@ -129,10 +169,10 @@ try:
     event_loop.run_forever()
 except KeyboardInterrupt:
     if verbose:
-        print('Keyboard interrupt\r\n')
+        LOG.info('Keyboard interrupt')
 finally:
     if verbose:
-        print('Closing event loop\r\n')
+        LOG.info('Closing event loop')
     btctrl.stop_scan_request()
     command = aiobs.HCI_Cmd_LE_Advertise(enable=False)
     btctrl.send_command(command)